+++ lang/src/org/partiql/lang/planner/transforms/AstToLogicalVisitorTransform.kt
@@ -294,9 +294,19 @@ internal class AstToLogicalVisitorTransform(
                             } else -> TODO("Only `DO REPLACE EXCLUDED` is supported in logical plan at the moment.")
                         }
                     }
-                    is PartiqlAst.ConflictAction.DoUpdate -> TODO(
-                        "`ON CONFLICT DO UPDATE` is not supported in logical plan yet."
-                    )
+                    is PartiqlAst.ConflictAction.DoUpdate -> {
+                        when (conflictAction.value) {
+                            PartiqlAst.OnConflictValue.Excluded() -> PartiqlLogical.build {
+                                dml(
+                                    target = dmlOp.target.toDmlTargetId(),
+                                    operation = dmlUpdate(),
+                                    rows = transformExpr(dmlOp.values),
+                                    metas = node.metas
+                                )
+                            }
+                            else -> TODO("Only `DO UPDATE EXCLUDED` is supported in logical plan at the moment.")
+                        }
+                    }
                     is PartiqlAst.ConflictAction.DoNothing -> TODO(
                         "`ON CONFLICT DO NOTHING` is not supported in logical plan yet."
                     )