+++ lang/src/org/partiql/lang/planner/transforms/LogicalResolvedToDefaultPhysicalVisitorTransform.kt
@@ -153,8 +153,9 @@ internal class LogicalResolvedToDefaultPhysicalVisitorTransform(
         val action = when (node.operation) {
             is PartiqlLogicalResolved.DmlOperation.DmlInsert -> DmlAction.INSERT
             is PartiqlLogicalResolved.DmlOperation.DmlDelete -> DmlAction.DELETE
-            is PartiqlLogicalResolved.DmlOperation.DmlReplace ->
-                TODO("DmlReplace physical transform hasn't been implemented yet")
+            is PartiqlLogicalResolved.DmlOperation.DmlReplace -> DmlAction.REPLACE
+            is PartiqlLogicalResolved.DmlOperation.DmlUpdate ->
+                TODO("DmlUpdate physical transform is not supported yet")
         }.name.toLowerCase()
 
         return PartiqlPhysical.build {