+++ lang/test/org/partiql/lang/eval/evaluatortestframework/PartiQLCompilerPipelineFactory.kt
@@ -93,7 +93,8 @@ internal class PartiQLCompilerPipelineFactory : PipelineFactory {
                 val statement = pipeline.compile(query)
                 return when (val result = statement.eval(session)) {
                     is PartiQLResult.Delete,
-                    is PartiQLResult.Insert -> error("DML is not supported by test suite")
+                    is PartiQLResult.Insert,
+                    is PartiQLResult.Replace -> error("DML is not supported by test suite")
                     is PartiQLResult.Value -> result.value
                 }
             }